For CHP applications
Assuming constant efficiencies (electrical and thermal), a CHP efficiency of 0.8 that  is constant when the engine is working at a load of 80 to 100%.
It was implmented an equation to determine the thermal energy available, the flow of fuel, and a restriction that can limit the amount of fuel available. It was also included a thermal demand.